Add edit function to created task items
This commit closes #8

taskItem was refactored to make the task title and description its own
component and add an edit button.

taskForm was refactored to take in arguments to display the task information
a user is editing and change the text button to "save" or "add task"
depending if the user is editing or adding a new task, respectively.

Many of the methods in taskController were refactored. Some new methods
were created as well. These methods now track and reference the task items
being edited through the use of the taskItem data attribute and references
to the taskForm that was targeted in the AddEditButtonListener method. in
addition, an unused form validation method (isFormTaskHeaderEmpty) was
removed since missingValueAggressiveValidation replaced it in a previous
commit.