Fix bug where canceling task edit saves changes
Fixes #13

The cancelTaskEditListener and AddEditButtonListener methods in taskController.js
were updated to use data- attributes to store the previous task header and
description data. On canceling an edit action, the task header and description
data would be pulled from the data- attributes.

Previously, canceling an edit action would cause taskController to pull
the task header and description data from the taskForm the user interacted
with.