Update to read and write from local storage
Closes #21

The cancel and edit listeners were refactored to use the
methods from webStorageController read and write to from
local storage on the browser. Task data is no longer stored
in a 'data-' attribute. This cleans up the code and makes it
easier to determine where the data is stored to and read from.

In addition, components were refactored to accept a task item object
of the task data making it easier to handle the data.