Replace taskItemNumber 'data-' attribute
Fixes #23

This commit replaces the taskItemNumber 'data-' attribute with taskItemId.
Previously, taskItemNumber was used to ID a task and held a number
(starting from 0) that determined based on the number of tasks created.
taskItemId replaces it with a more unique ID based on the due date of the
task and how many tasks were created on that day. It takes the format of
YYYYMMDD[task #] where [task #] is replaced with a number starting from 0
(e.g. 20221122 to indicate the due date is 22NOV2022 and it's the first
task for that due date).