Refactor taskID
Fixes #25

Task ID was previously generated based on the due date and task #. Task ID
is now generated based on a random 8 digit integer to keep the ID unique
and independent of the data for a given task. A random unique ID was used
mainly to keep the task ID unique. It can help with obfuscating the task
associated with the ID but that was not the main goal for implementing the
ID system.