Add Inbox sort method and refactor code
Closes #27

In order to implement the Inbox sorting method, many methods were refactored
and new components/ methods were created. The Today and Upcoming menu tab
sorting methods were refactored due to the data structure change in localStorage,
which is discussed more below. The taskPriorityBoard component was created to
visually order the tasks by priority for the user. Each taskPriorityBoard has
it's own addNewTaskButton to create and add the task to the chosen priority board.
The sorting methods were also moved from the menuController file to the
taskController file since the sort methods deal more with tasks than the
menu.

The webStorageController methods were refactored or created due to the localStorage
data structure change as well. There are now general and task-specific methods
to get and set data to and from localStorage.

mainContentHeader and taskViewer were updated to prep for a future update
where the Inbox menu tab is loaded first on page load.

In addition, the data structure of
localStorage was refactored to make it clear what the data is and pre-sort
the tasks by priority. Tasks are pre-sorted by task priority because on
page load the application will default to the Inbox menu tab. Having the data
pre-sorted by priority will help with speed and resource usage on first load since
the tasks don't need to be sorted as much on load. Index.js was updated to
initialize localStorage with the new data structure on page load if it doesn't
exist.

Minor styling was added for the taskPriorityBoard components.