Add feature to load tasks by priority for Inbox
Closes #30

On page load, the app should display the Inbox content and sort any saved
tasks by priority. To implement this, taskViewer will render as an empty
container. The task priority boards will be added right after and filled
with any saved tasks in localStorage. In addition, previousMenuTab in localStorage
is initialized with the value "Inbox" in order to prevent resorting/ displaying
the same content if the user clicks on the Inbox menu tab.