Refactor app to only have one add task button
Closes #31

The app was refactored to only have one add task button to
simplify the UI and make the code base less complex when it
comes to adding, editing, or canceling tasks. Before, the number
of add task buttons were dependent on the task view (Inbox, Today, etc.)
that was selected. This made managing tasks and the UI more difficult
and cluttered.

The only add task button on the app now resides in the header of the page
which makes it easier and more intuitive to locate. When adding or
editing a task, the form now appears as an overlay over the entire app
to make it easier to locate the form. Prior to this change, the form
would be added in the flow of the tasks making it difficult to locate.

New methods were created as well to help manage individual tasks. Many of the
methods were renamed as well to help clarify what they do or are.

Some styling changes were made to make viewing the changes more
appealing.