Add secondary date task boards for upcoming tasks
Fixes #32

Refactored taskPriorityBoard component into secondaryTaskBoard to make
the component reusable. secondaryTaskBoard was used to create the date
task boards for the Upcoming task view. The sorting functions for
individual and all tasks were refactored to account for the new
date task boards.

In addition, a bug was fixed where if the duedate was eight days away from the current day,
function differenceInDays would round the difference up causing any tasks on the
eighth day to be added to the seventh day task board. This is due to
task due dates not having a time component while today's date does.