Fix task duplicating and not being shown on edit
Fixes #35

When saving an edited task, the old and newly edited versions of the
tasks were not removed from localStorage. A new function in webStorageController
was added to remove the old task data and fixes the issue.

This also fixes the issue of the new task not being added to the DOM.