Refactor UI style fix scrolling bugs
Fixes #33 and fixes #34

The app was restyled so that the entire app is contained within the size
of the viewport. Any overflow that occurs outside of <div class="main-content>
is hidden to prevent the entire page from scrolling. Only
<div class="main-content> will be scrollable.

This also fixes a bug where users can still scroll the page when an
overlay form is shown. Limiting the app to the size of the viewport by
default prevents the user from scrolling the page an further.