Fix task being duplicated on edit cancel
Fixes #36

Function cancelTaskFormListener was refactored to remove code that
duplicated the task in the DOM. The code was overlooked when refactoring
the app to have a single add new task button in a prior commit. Reference
issue #31. The function now only removes the form overlay container. This
clears the form on removal so there is no remaining data in the form that
could cause unintended side effects.