Add feature to delete tasks
Closes #37

A delete button with an icon was added to each task. Clicking the button
will remove the task from the DOM and localStorage data. The deleteTaskItem
function was refactored to avoid having to grab the task data first and
determine the task's priority key before
locating the task to delete it. Instead, delete task will determine if the task
exists and delete it if found. This simplifies and speeds up the search process.