Add a feature to remove and store completed tasks
Closes #41

Completed tasks are removed from the DOM and the taskData key in localStorage, and
copied over to the completed key. There the tasks are grouped based on the
day the task was completed. A new function was created in webStorageController
to specifically handle completed tasks. toggleTaskStatus in taskController was
refactored to setTaskCompleted to have a dedicated function for removing the tasks
from the taskData key, copying the tasks to the completed key, and removing the
task from the DOM.