Add feature to sort and display completed tasks
Closes #29

When viewing completed tasks, the app displays the completed tasks from
the most recent date to the oldest date (descending order). Displayed
tasks are visually crossed out and the check mark is visible to help
indicate that the tasks are completed. The date key for variable
todaysDate in function saveTaskToCompleted was updated so that new Date() and
date-fns functions can recognize/ process the string as a date.