Add feature to "uncompleted" completed tasks
Closes #42

Added a feature to give users the ability to undo task completions by
checking the checkbox. Once checked, tasks are removed from the completed
key and placed back into taskData with the task's original data.

A new helper function was created in webStorageController to help determine
if a task should be removed, saved, or edited based on the current task
view.

saveTaskItem, getTaskItem, and deleteTaskItem were refactored so that items
could be processed based on the localStorage key provided. This helps make
the code more modular and prepares it for a future update for project
related tasks.

A small bug for insertTaskForTodayView in taskController was also fixed
while testing the feature update where variable taskDueDate was being
assigned a task priority key instead of the dueDateValue.