Render project tabs on page load
Fixes #47

On page load, project names from the project key in localStorage are
rendered as tabs on the menu bar under the Project section.

As part of this update, addProjectTabListener was refactored to fix an
issue where the event listener was looking for a project tab that has not
rendered on the DOM yet. It now adds the event listener to the project tab
before the DOM is created. addNewProjectButtonListener was refactored due to
the changes in addProjectTabListener.