Add function to add new tasks to projects
Closes #44

Many of the functions in webStorageController were refactored to update
function or variable names to clarify what they do or hold and allow new
tasks to be added to a specific project.

Instead of creating having a separate save function for the different types
of save locations (taskData, completed, and projects) saveTaskItem was
refactored and repurposed as the only save function. Instead, the code needed
to save data to specific locations was split up. This was done to reduce
having to import functions from webStorageController to different files
and make it easier to maintain. Originally, I had thought creating
separate saves files would be better but changed my mind once I started
realizing having multiple save functions adds more complexity in determining
which one to use and how the code base is connected together.