Show, sort, and group project tasks
Closes #48

Project tasks are grouped by due dates and sorted based on the due dates
from the earliest to the furthest date. This is done to help the user
visually see what tasks need to be done first. An "overdue" label is
added to each date board to further help determine which tasks need to be
completed first.

In addition, some of the functions related to sort tasks were renamed
to help clarify what they do. Some were refactored in order to implement
the feature added in this commit.