Refactor how new project tasks displayed and added
Fixes #49

A new function insertTaskForProject which clears the primary task
board and calls sortTasksByProject was created. Instead of trying to
determine where to insert a new project task on the DOM, clearing and resorting
all project tasks was chosen. Attempting to determine and sort
where a new task should go caused noticeable application lag. Clearing and
resorting the project tasks was much quicker. This due to the structure of
app components and how the content is displayed.

The other individual task insertion functions will be kept for now
since they have no noticeable effects on the app performance.