Delete secondary task board when last task deleted
Closes #51

When deleting the last task for a given secondary task board for the
inbox, completed, and project view, the secondary task board is removed
as well. This is declutter the UI for those views. The today and upcoming
views require the secondary task boards to stay in order to convey
if any tasks are remaining for a given day.