Add a label for tasks
Closes #52

Tasks now are labeled according to where they were created. There are two
types of labels: general and project task. The goal of this is visually
indicate to the user when looking at a task in the inbox, today,
upcoming, and completed task views what kind of task it is. This also
helps with issue #53 in determining where a task should be placed back in.