Fix logo and add overlay closing by clicking it
In Firefox, the logo text would wrap and stick outside of it's parent
container. This was due to the browser applying flex shrink to the text.

A feature was also added to allow users to close any overlays/ modals by
clicking on the overlay. This was added to provide a more intuitive way to
close them.